Removed use of module.cc in TrackingTools/Producers
The declaration of the producer classes and the registration of
the plugin were moved to the .cc files.
