+++ TrackingTools/Producers/src/StraightLinePropagatorESProducer.cc
@@ -1,14 +1,27 @@
-#include "TrackingTools/Producers/interface/StraightLinePropagatorESProducer.h"
+#include "FWCore/Framework/interface/ESProducer.h"
+#include "FWCore/ParameterSet/interface/ParameterSet.h"
+#include "TrackingTools/GeomPropagators/interface/StraightLinePropagator.h"
+#include "TrackingTools/Records/interface/TrackingComponentsRecord.h"
 #include "MagneticField/Engine/interface/MagneticField.h"
 #include "MagneticField/Records/interface/IdealMagneticFieldRecord.h"
 
 #include "FWCore/Framework/interface/EventSetup.h"
 #include "FWCore/Framework/interface/ESHandle.h"
 #include "FWCore/Framework/interface/ModuleFactory.h"
-#include "FWCore/Framework/interface/ESProducer.h"
 
 #include <string>
 #include <memory>
+
+class StraightLinePropagatorESProducer : public edm::ESProducer {
+public:
+  StraightLinePropagatorESProducer(const edm::ParameterSet &p);
+  ~StraightLinePropagatorESProducer() override;
+  std::unique_ptr<Propagator> produce(const TrackingComponentsRecord &);
+
+private:
+  edm::ParameterSet pset_;
+};
+
 
 using namespace edm;
 
@@ -37,3 +50,5 @@ std::unique_ptr<Propagator> StraightLinePropagatorESProducer::produce(const Trac
     dir = anyDirection;
   return std::make_unique<StraightLinePropagator>(&(*magfield), dir);
 }
+
+DEFINE_FWK_EVENTSETUP_MODULE(StraightLinePropagatorESProducer);
