+++ TrackingTools/Producers/src/TrajectoryCleanerESProducer.cc
@@ -1,5 +1,28 @@
-#include "TrackingTools/Producers/interface/TrajectoryCleanerESProducer.h"
+#include "FWCore/Framework/interface/ModuleFactory.h"
+#include "FWCore/Framework/interface/ESProducer.h"
+
+#include "FWCore/Framework/interface/ESHandle.h"
+
+#include "TrackingTools/Records/interface/TrackingComponentsRecord.h"
+#include "TrackingTools/TrajectoryCleaning/interface/TrajectoryCleaner.h"
+
+
 #include "TrackingTools/TrajectoryCleaning/interface/TrajectoryCleanerFactory.h"
+
+class TrajectoryCleanerESProducer : public edm::ESProducer {
+public:
+  TrajectoryCleanerESProducer(const edm::ParameterSet&);
+  ~TrajectoryCleanerESProducer() override;
+
+  typedef std::unique_ptr<TrajectoryCleaner> ReturnType;
+
+  ReturnType produce(const TrackingComponentsRecord&);
+
+private:
+  std::string theComponentName;
+  std::string theComponentType;
+  edm::ParameterSet theConfig;
+};
 
 TrajectoryCleanerESProducer::TrajectoryCleanerESProducer(const edm::ParameterSet& iConfig) {
   theComponentName = iConfig.getParameter<std::string>("ComponentName");
@@ -17,3 +40,5 @@ TrajectoryCleanerESProducer::ReturnType TrajectoryCleanerESProducer::produce(con
 
   return ReturnType(TrajectoryCleanerFactory::get()->create(theComponentType, theConfig));
 }
+
+DEFINE_FWK_EVENTSETUP_MODULE(TrajectoryCleanerESProducer);
